small bug fixes
