+++ frontend/stylesheets/quizManagement.css
@@ -189,24 +189,30 @@
 .quiz-view-details-modal ul li {
   font-size: 1rem;
 }
+#quiz-manage-container {
+  min-height: 70vh;
+}
 /*  Large devices (desktops, less than 1200px) */
 @media (max-width: 1199.98px) {
   .quiz-info {
     flex-direction: column;
   }
 }
-
-/*  Small devices (landscape phones, less than 768px) */
-@media (max-width: 767.98px) {
+/* Medium devices (tablets, less than 992px) */
+@media (max-width: 991.98px) {
   .btns-wrapper {
     flex-direction: column;
+    width: 100%;
   }
+  .btns-wrapper button {
+    width: 100%;
+  }
+}
+/*  Small devices (landscape phones, less than 768px) */
+@media (max-width: 767.98px) {
   .quiz-management form {
     flex-direction: column;
     height: fit-content;
-  }
-  .btns-wrapper {
-    width: 100%;
   }
   .quiz button {
     width: 100%;
