+++ frontend/stylesheets/quizSearch.css
@@ -76,6 +76,7 @@
   padding: 1rem 0;
   display: flex;
   flex-direction: column;
+  width: 100%;
   gap: 10px;
 }
 .quiz-banner img {
