+++ frontend/stylesheets/userManagement.css
@@ -112,7 +112,15 @@
     flex-direction: column;
   }
 }
-
+@media (max-width: 991.98px) {
+  .btns-wrapper {
+    flex-direction: column;
+    width: 100%;
+  }
+  .btns-wrapper button {
+    width: 100%;
+  }
+}
 /*  Small devices (landscape phones, less than 768px) */
 @media (max-width: 767.98px) {
   .btns-wrapper {
