+++ rest/dao/HistoryDao.class.php
@@ -36,4 +36,53 @@ class HistoryDao extends BaseDao {
       return $quizHistoryData;
     }
   }
+  public function insertHistory($payload) {
+    $quizInfo = $payload["takenQuiz"];
+
+    $query = "INSERT INTO quiz_history (quiz_id, user_id, timeTaken, correctAnswers)
+              VALUES (:quiz_id, :user_id, :timeTaken, :correctAnswers)";
+        
+    $params = array (
+        ':quiz_id' => $quizInfo["quizID"],
+        ':user_id' => $quizInfo["userID"],
+        ':timeTaken' => $quizInfo["timeTaken"],
+        ':correctAnswers' => $quizInfo["correctAnswers"],
+    );
+    $this->execute($query, $params);
+    $quizHistoryID = $this->connection->lastInsertId(); 
+
+    foreach ($quizInfo["answers"] as $answer) {
+        $query2 = "INSERT INTO quiz_answer (quiz_history_id)
+                   VALUES (:quiz_history_id)";
+        
+        $params2 = array (
+            ':quiz_history_id' => $quizHistoryID,
+        );
+        $this->execute($query2, $params2);
+        $quiz_answer = $this->connection->lastInsertId();
+
+        if (isset($answer["userAnswer"]) && is_array($answer["userAnswer"])) {
+            foreach($answer["userAnswer"] as $field) {
+                $query3 = "INSERT INTO answer (quiz_answer_id, title, isCorrect)
+                           VALUES (:quiz_answer_id, :title, :isCorrect)";
+        
+                $params3 = array (
+                    ':quiz_answer_id' => $quiz_answer,
+                    ':title' => $field["title"],
+                    ':isCorrect' => $field["isCorrect"] == "true" ? 1 : 0, 
+                );
+                $this->execute($query3, $params3);
+            }
+        } else {
+          $query3 = "INSERT INTO answer (quiz_answer_id, title, isCorrect)
+          VALUES (:quiz_answer_id, :title, :isCorrect)";
+          $params3 = array (
+              ':quiz_answer_id' => $quiz_answer,
+              ':title' => null,
+              ':isCorrect' => 0, 
+            );
+          $this->execute($query3, $params3);
+        }
+    }
+  }
 }
