+++ rest/routes/postQuizHistory.php
@@ -1,13 +1,11 @@
 <?php 
 
-require '../services/UserService.class.php';
-
+require '../services/HistoryService.class.php';
 
 if(isset($_POST["takenQuiz"])) {
-  $userHistory = new UserService();
-  $result = $userHistory->insertHistory($_POST);
+  $history = new HistoryService();
   header('HTTP/1.1 200 OK');
-  echo json_encode($result);
+  echo json_encode($history->insertHistory($_POST));
 }  else {
   header('HTTP/1.1 400 Bad Request');
   echo json_encode(array('error' => 'No quiz info provided'));
