+++ rest/services/HistoryService.class.php
@@ -2,16 +2,19 @@
 
 require_once __DIR__ . '/../dao/HistoryDao.class.php';
 
-class QuizService {
+class HistoryService {
   private $historyDao;
 
   public function __construct() {
     $this->historyDao = new HistoryDao();
   }
   public function getQuizHistory($id) {
-    return $this->getQuizHistory($id);
+    return $this->historyDao->getQuizHistory($id);
   }
   public function getQuizHistoryByID($email, $quiz_id) {
-    return $this->getQuizHistoryByID($email, $quiz_id); 
+    return $this->historyDao->getQuizHistoryByID($email, $quiz_id); 
+  }
+  public function insertHistory($payload) {
+    return $this->historyDao->insertHistory($payload);
   }
 }
