+++ rest/services/UserService.class.php
@@ -29,7 +29,4 @@ class UserService {
   public function getLeaderboard() {
     return $this->userDao->getLeaderboard();
   }
-  public function insertHistory($payload) {
-    return $this->userDao->insertHistory($payload);
-  }
 }
